+++ example/android/app/src/main/kotlin/example/decoration/corner/rotated/example/MainActivity.kt
@@ -1,0 +1,6 @@
+package example.decoration.corner.rotated.example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}