+++ lib/src/badge_position.dart
@@ -1,0 +1,1 @@
+enum BadgePosition { topStart, topEnd, bottomStart, bottomEnd }