+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+## [2.1.0] - 12.03.2021
+
+* Increased minimum SDK version to 2.17.0
+* Breaking changes: 
+- now there are two named constructors: ```withColor()``` and ```withGradient()```
+- ```BadgeGeometry``` replaced by three parameters: ```badgeSize```, ```badgeCornerRadius``` and ```badgePosition```
+- ```LabelInsets``` replaces by two parameters: ```spanBaselineShift``` and ```spanHorizontalOffset```
+
+* Example code refactored 
+* added experimental feature ```isEmoji: true``` it disables Span rotation
+
 ## [2.0.0] - 26.03.2021
 
 * Migrate to null safety.